# Optimizing client APIs pathspec, parent and path_components to be O(1)