Overall, this prevents recursive construction of parents. However, lazy
construction of parents does incur a one time cost (see perf metrics below).

* Since objects can either be created using pathspec or parent;
compute pathspec using parent_pathspec when absent.
There will be no cyclic dependency in calls.

* Construct parent object from pathspec when absent and cache it.

**Testing**
* Local tests `python3 ./run_tests.py --debug --contexts python3-all-local`
* Ran through the tutorial using this PR branch


**Performance Test**

I wrote a small script to measure performance (let me know if there is interest in this script, I can clean it up and create a new PR) for some of the client APIs. These were measured in local model. Summary below:

| API  (100 iterations)     | Latency(ms) for Baseline | Latency(ms) for this PR |
| ----------- | ----------- | ----------- |
|   Flow creation `Flow()`   | 934.5665979385376  |  846.9134831428528 |
|  `flow.path_components`   | 0.001070499420166015  |  0.0010228157043457031 |
|  `flow.pathspec`   | 0.002040863037109375  |  0.00044345855712890625 |
|  `flow.parent`   | 0.0004553794860839844  |  0.00048160552978515625 |
|  Run creation `Run()`   | 1341.3884568214417 |  398.44561100006104 |
|  `run.path_components`   | 0.0011157989501953125  |  0.0010657310485839844 |
|  `run.pathspec`   | 0.00213623046875  |  0.0004863739013671875 |
|  `run.parent`   | 0.0003552436828613281  |  4.708406925201416 |
|  Step creation `Step()`   | 1697.720820903778  |  389.5357656478882 |
|  `step.path_components`   | 0.001125335693359375  |  0.0011301040649414062 |
|  `step.pathspec`   | 0.0013375282287597656  |  0.0004792213439941406 |
|  `step.parent`   | 0.0002503395080566406  |  4.700286388397217 |